+++ package.json
@@ -54,6 +54,7 @@
     "eslint-plugin-import": "^2.27.5",
     "prop-types": "^15.8.1",
     "storybook": "^7.0.9",
+    "storybook-addon-recoil-flow": "^1.5.1",
     "tsconfig-paths-webpack-plugin": "^4.0.1",
     "webpack": "^5.82.0"
   }
