+++ src/components/common/Header/index.tsx
@@ -1,5 +1,5 @@
-import { HEADER_LOGO } from '@assets';
 import { styled } from 'styled-components';
+import { HEADER_LOGO } from '@assets';
 import { theme } from '@styles/theme';
 
 const Header = () => {
